Exclude unit tests from upstream that are known to fail in Brave
This patch excludes entire *_unittests.cc files from being compiled so
that we avoid running any test that we know are going to fail because
of the many ways in which Brave is different than Chromium upstream.

Note that it's possible that this way of excluding tests is a bit too
agressive (i.e. it's likely that some unit tests inside the excluded
files would pass), but for now it's a good enough initial approach as
it enables us to considerably increase test coverage without having
to maintain patches in a too intrusive way.

As a reference, a this time of this patch's writing (on top of Brave
1.22.12 / Chromium 89.0.4389.40), this are the stats when running
upstream's unit tests on a Linux/Debug build, without this patch:

  * Total run: 13486 tests
  * Passed: 11625 tests
  * Not passed: 1861 tests
    - Failed: 168 tests
    - Crashed: 1693 tests
    - Timed out: 0 tests

With this patch applied, the numbers look like this:

  * Total run: 10057 tests
  * Passed: 10057 tests
  * Not passed: 0 tests

That is, what we have with this patch applied looks as follows:

  * 10057/13486 -> 74.57% of ALL the original tests being run
  * 10057/11625 -> 86.51% of the PASSING TESTS still being run

In other words, we're increasing test coverage in 10057 unit tests
in a relatively clean way (i.e. no complex patching) while "only"
losing 13.49% of the unit tests that would run and pass if we were
not excluding them in such an agressive way.

Fix https://github.com/brave/brave-browser/issues/8376