Exclude browser tests from upstream that are known to fail in Brave
This patch excludes entire *_browsertest.cc files from being compiled
so that we avoid running any test that we know are going to fail due
to the many ways in which Brave is different than Chromium upstream.

Note that it's possible that this way of excluding tests is a bit too
agressive (i.e. it's likely that some browser tests inside the excluded
files would pass), but for now it's a good enough initial approach as
it enables us to considerably increase test coverage without having
to maintain patches in a too intrusive way.

As a reference, a this time of this patch's writing (on top of Brave
1.22.18 / Chromium 89.0.4389.40), these are the stats when running
upstream's browser tests on a Linux/Debug build, without this patch:

  * Total run: 13337 tests
  * Passed: 11429 tests
  * Not passed: 1908 tests
    - Failed: 871 tests
    - Crashed: 555 tests
    - Timed out: 482 tests

With this patch applied, the numbers look like this:

  * Total run: 6084 tests
  * Passed: 6084 tests
  * Not passed: 0 tests

That is, what we have with this patch applied looks as follows:

  * 6084/13337 -> 45.62% of ALL the original tests being run
  * 6084/11429 -> 53.23% of the PASSING TESTS still being run

In other words, we're increasing test coverage in 6084 browser tests
in a relatively clean way (i.e. no complex patching) while losing
46.77% of the browser tests that would run and pass if we were not
excluding them in such an agressive way. Not as good results as for
the case of unit tests, but still an improvement for test coverage.

Fix https://github.com/brave/brave-browser/issues/8297