Adapt launch of unsandboxed processes via SandboxWin::StartSandboxedProcess()
The place we need to patch has now moved from there to the private helper
function LaunchWithoutSandbox(), where we don't have access to the string
with the process type. Therefore, we need to adapt slightly the helper
in the chromium_src override and move the place it's called from.

Chromium change:
https://chromium.googlesource.com/chromium/src.git/+/92757d2ff56133de2a8999d8327d5d108f632565

commit 92757d2ff56133de2a8999d8327d5d108f632565
Author: Alex Gough <ajgo@chromium.org>
Date:   Thu Feb 11 01:02:37 2021 +0000

    Refactor unsandboxed process launching.

    No functional changes. I intend to expand this helper function to cope
    with CETCOMPAT. This refactor makes it clearer that this path is taken
    and does not continue into the rest of the sandboxed flow.

    Bug: 1174399